Refactor devices: Separate decryption from parsing
This will allow for simpler tests without needing decryption keys.
